requirements.txt
aiogram
python-dotenv
black